# docs/swagger.yaml
definitions:
  big.Int:
    type: object
...
  db.CreateDestinationParams:
    properties:
      anh:
        type: string
      khu_vuc:
        type: string
      kinh_do:
        $ref: '#/definitions/pgtype.Numeric'
      mo_ta:
        type: string
      quoc_gia:
        type: string
      ten:
        type: string
      vi_do:
        $ref: '#/definitions/pgtype.Numeric'
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Numeric:
    properties:
      exp:
        format: int32
        type: integer
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      int:
        $ref: '#/definitions/big.Int'
      naN:
        type: boolean
      valid:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Travia API for travel management
  termsOfService: http://swagger.io/terms/
  title: Travia API
  version: "1.0"
paths:
...
  /destination:
    post:
      consumes:
      - application/json
      description: Tạo điểm đến
      parameters:
      - description: Destination
        in: body
        name: destination
        required: true
        schema:
          $ref: '#/definitions/db.CreateDestinationParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.CreateDestinationParams'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Tạo điểm đến
      tags:
      - Destination
  /destination/city/{province}:
    get:
      consumes:
      - application/json
      description: Lấy danh sách thành phố theo tỉnh thành
      parameters:
      - description: Province
        in: path
        name: province
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Lấy danh sách thành phố theo tỉnh thành
      tags:
      - Destination
  /destination/country:
    get:
      consumes:
      - application/json
      description: Lấy danh sách quốc gia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Lấy danh sách quốc gia
      tags:
      - Destination
  /destination/province/{country}:
    get:
      consumes:
      - application/json
      description: Lấy danh sách tỉnh thành theo quốc gia
      parameters:
      - description: Country
        in: path
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Lấy danh sách tỉnh thành theo quốc gia
      tags:
      - Destination
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"